+++ packages/base/src/index.ts
@@ -1,9 +1,10 @@
 export * from './commands';
-export * from './formdialog';
+export * from './dialogs/formdialog';
 export * from './mainview';
 export * from './tools';
 export * from './types';
 export * from './widget';
+export * from './formbuilder';
 
 export * from './panelview';
 export * from './toolbar';
